--- src/Pages/CreateAccount.tsx
@@ -1,12 +1,0 @@
-import * as React from "react";
-import signUp from "../API/signup";
-
-const CreateAccount = () => {
-  const onClick = async () => {
-    const [token, error] = await signUp("helldddo@hello.com", "hellohello");
-    console.log(token, error);
-  };
-  return <div onClick={onClick}>CreateAccount</div>;
-};
-
-export default CreateAccount;
